# Optimized Tokenize function