It's more like an V2.5, but Go doesn't allow dots within the function names ... so here we are with an V3. Feel free to use these small changes into the existing V2 ... but for better showing of the performance increase, I left it in for now.

```bash
~ go test ./... -bench=. -test.benchmem
goos: linux
goarch: amd64
pkg: github.com/emetriq/gourltokenizer/tokenizer
cpu: AMD Ryzen 7 3700X 8-Core Processor             
BenchmarkEscapedURLTokenizerV3-16        1271954               935.8 ns/op           464 B/op          3 allocs/op
BenchmarkURLTokenizerV3-16               2607987               457.4 ns/op           256 B/op          1 allocs/op
BenchmarkURLTokenizerV3Fast-16           4605972               248.3 ns/op           256 B/op          1 allocs/op
BenchmarkEscapedURLTokenizerV2-16        1234116               959.0 ns/op           464 B/op          3 allocs/op
BenchmarkURLTokenizerV2-16               2165749               552.2 ns/op           256 B/op          1 allocs/op
BenchmarkURLTokenizerV2Fast-16           4374403               272.5 ns/op           256 B/op          1 allocs/op
BenchmarkEscapedURLTokenizerV1-16        1246105               963.0 ns/op           464 B/op          3 allocs/op
BenchmarkURLTokenizerV1-16               1668098               718.9 ns/op           272 B/op          2 allocs/op
BenchmarkTokenizerV1-16                  6767202               169.8 ns/op           256 B/op          1 allocs/op
BenchmarkTokenizerV2-16                  5556195               214.9 ns/op           256 B/op          1 allocs/op
BenchmarkTokenizerV3-16                  6243402               186.7 ns/op           256 B/op          1 allocs/op
PASS
ok      github.com/emetriq/gourltokenizer/tokenizer     18.833s

~ go test ./...
ok      github.com/emetriq/gourltokenizer/tokenizer     0.002s
```

Added an [..]Escaped[..] benchmark, because most performance profits come up, if the given URL doesn't contain any escaped parts. Apart from that, using a byte instead of a rune gives us some performance increase, but results in only accepting the ASCII chars... but since it's already filtering out anything else than a-z, it shouldn't have an impact.